requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
pandas
Pillow
tqdm
pyyaml